Move user data store folder
