+++ packages/data/src/index.js
@@ -9,9 +9,9 @@ export { withPluginsHydration } from './plugins/with-plugins-hydration';
 export { ONBOARDING_STORE_NAME } from './onboarding';
 export { withOnboardingHydration } from './onboarding/with-onboarding-hydration';
 
-export { USER_STORE_NAME } from './user-preferences';
-export { withCurrentUserHydration } from './user-preferences/with-current-user-hydration';
-export { useUserPreferences } from './user-preferences/use-user-preferences';
+export { USER_STORE_NAME } from './user';
+export { withCurrentUserHydration } from './user/with-current-user-hydration';
+export { useUserPreferences } from './user/use-user-preferences';
 
 export { OPTIONS_STORE_NAME } from './options';
 export { withOptionsHydration } from './options/with-options-hydration';
